requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
pandas
requests
